Add record count to victims database interface.
Currently there is no way for a client to determine the number of
records in the victims database. One thing that this is useful for
is determining if a scan is worthwhile or if an update should be issued.
The only other workaround is to check the lastUpdate vs time(0) but this is
not derived from the actual database state.